pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "customer-churn-retrieval"
version = "0.2.0"
description = "Hybrid retrieval-based customer churn prediction pipeline"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "OpenAI ChatGPT" }
]
dependencies = [
  "chromadb>=0.5.0",
  "langchain-community>=0.2.0",
  "langchain-openai>=0.1.0",
  "matplotlib>=3.8.0",
  "natsort>=8.4.0",
  "numpy>=1.26.0",
  "pandas>=2.1.0",
  "python-dotenv>=1.0.0",
  "scikit-learn>=1.4.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0"
]

[project.scripts]
churn-pipeline = "churn_retrieval.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
